Write detect_cycles as a function of Module
The different algorithm we use, currently DFS and Tarjan, may have
different typing requirements.  This lets us keep them separate.
For example, the Tarjan does not require comparability.

This also lets us keep everything in `project.main` expressed in the
domain (that is, as a function of `Module`) while not sacrificing
generalization in the implementation of the algorithms.