Fix main() and entry point
`argparse` does not need `argv`, quoting the documentation:

> In a script, `parse_args()` will typically be called with
> no arguments [...]

This lets us make an argument-less `main()` function that
can be used easily as the entry point in setup.py.

With this change, installing the source dist provides a
script that can be called simply with

```
py_import_cycles --help
```

or any relevant arguments.